Fix middleware location for Next.js
